few changes read below
1) deleted int casting in calc
2) created getEndOfExpression
3) in conditionParser casted to float ill explane
4) filtered "" in the lexer
5)changed calc to work to the end of the expression only.

+++ src/client_side/interpreter/DefaultConditionParser.java
@@ -14,14 +14,14 @@ public class DefaultConditionParser implements ConditionParser {
 
     @Override
     public Boolean parse(List<String> tokens, int startIndex) throws CannotInterpretException {
-        double operand1, operand2;
+        float operand1, operand2;
         Parser mp = new Parser(symbolTable);
         int operand1EndPos;
         try {
-            operand1 = mp.calc(tokens, startIndex);
+            operand1 = (float) mp.calc(tokens, startIndex);
             operand1EndPos = mp.getEndOfExpression(tokens, startIndex);
 
-            operand2 = mp.calc(tokens, operand1EndPos + /*operatorSize*/1 + 1);
+            operand2 = (float) mp.calc(tokens, operand1EndPos + /*operatorSize*/1 + 1);
         } catch (NumberFormatException e) {
             throw new CannotInterpretException("Cannot resolve condition operands", startIndex);
         }
